refactor: 💡 Migrate SQFormMaskedReadOnlyField to TypeScript
✅ Closes: #474
